@@ -10,21 +10,22 @@
     "eject": "react-scripts eject"
   },
   "dependencies": {
-    "axios": "^0.24.0",
-    "classnames": "^2.3.1",
+    "@leecheuk/react-google-login": "^5.4.1",
+    "axios": "^1.2.2",
+    "classnames": "^2.3.2",
     "gapi-script": "^1.2.0",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
     "react-bottom-scroll-listener": "^5.1.0",
-    "react-dom": "^17.0.2",
-    "react-google-login": "^5.2.2",
-    "react-hook-form": "^7.23.0",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0",
-    "redux": "^4.1.2",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
+    "react-hook-form": "^7.41.5",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.6.2",
+    "react-scripts": "^5.0.1",
+    "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.4.1"
+    "redux-thunk": "^2.4.2"
   },
   "eslintConfig": {
     "extends": [
