@@ -9,7 +9,7 @@ import { yupResolver } from '@hookform/resolvers/yup'
 import styles from "./styles.module.css"
 import AppBar from "../appbar/AppBar"
 import ImagePreviewer from "../image-previewer/ImagePreviewer"
-import MetaTags from 'react-meta-tags'
+import { Helmet } from "react-helmet";
 
 // TODO: валидация
 const schema = yup.object({
@@ -72,14 +72,14 @@ const EditProfile = ({user, profile, editProfile, getCurrentProfile}) => {
   }
 
   const comeBack = () => {
-    navigate(`/@${user.login}`)
+    navigate(`/${user.login}`)
   }
 
   return (
     <div>
-       <MetaTags>
+       <Helmet>
         {user && <title>Edit profile | Tell Opinion</title>}
-      </MetaTags>
+      </Helmet>
       <AppBar/>
       <main className={styles.container}>
         <header className={styles.header}>
