@@ -2,12 +2,12 @@ const mongoose = require('mongoose');
 const config = require('config');
 const db = config.get('mongoURI');
 
+mongoose.set('strictQuery', false);
+
 const connectDB = async () => {
 	try {
 		await mongoose.connect(db, {
 			useNewUrlParser: true,
-			useCreateIndex: true,
-			useFindAndModify: false,
 			useUnifiedTopology: true
 		});
 
